Fixing warnings in Windows build.

+++ modules/com/zorba-xquery/www/modules/datetime.xq.src/datetime.cpp
@@ -66,8 +66,14 @@ CurrentDateTimeFunction::evaluate(
 
   zorba::Item lDateTime = 
     Zorba::getInstance(0)->getItemFactory()->createDateTime(
-      (short)(gmtm.tm_year + 1900), (short)(gmtm.tm_mon + 1), (short)gmtm.tm_mday, (short)gmtm.tm_hour,
-      (short)gmtm.tm_min, gmtm.tm_sec + lMilliTm/1000.0, (short)(lTimezone/3600));
+      static_cast<short>(gmtm.tm_year + 1900),
+      static_cast<short>(gmtm.tm_mon + 1),
+      static_cast<short>(gmtm.tm_mday),
+      static_cast<short>(gmtm.tm_hour),
+      static_cast<short>(gmtm.tm_min),
+      gmtm.tm_sec + lMilliTm/1000.0,
+      static_cast<short>(lTimezone/3600)
+    );
 
   return zorba::ItemSequence_t(new zorba::SingletonItemSequence(lDateTime));
 }
@@ -88,7 +94,10 @@ CurrentDateFunction::evaluate(
 
   zorba::Item lDateTime = 
     Zorba::getInstance(0)->getItemFactory()->createDate(
-      gmtm.tm_year + 1900, gmtm.tm_mon + 1, gmtm.tm_mday);
+      static_cast<short>(gmtm.tm_year + 1900),
+      static_cast<short>(gmtm.tm_mon + 1),
+      static_cast<short>(gmtm.tm_mday)
+    );
 
   return zorba::ItemSequence_t(new zorba::SingletonItemSequence(lDateTime));
 }
@@ -109,7 +118,11 @@ CurrentTimeFunction::evaluate(
 
   zorba::Item lDateTime = 
     Zorba::getInstance(0)->getItemFactory()->createTime(
-      (short)gmtm.tm_hour, (short)gmtm.tm_min, gmtm.tm_sec + lMilliTm/1000.0, (short)(lTimezone/3600));
+      static_cast<short>(gmtm.tm_hour),
+      static_cast<short>(gmtm.tm_min),
+      gmtm.tm_sec + lMilliTm/1000.0,
+      static_cast<short>(lTimezone/3600)
+    );
 
   return zorba::ItemSequence_t(new zorba::SingletonItemSequence(lDateTime));
 }
